+++ backend/pom.xml
@@ -8,10 +8,10 @@
 		<version>3.1.0</version>
 		<relativePath/>
 	</parent>
-	<groupId>dev.jaczerob</groupId>
-	<artifactId>madamchuckle</artifactId>
+	<groupId>dev.jaczerob.madamchuckle</groupId>
+	<artifactId>server</artifactId>
 	<version>0.0.2</version>
-	<name>madamchuckle</name>
+	<name>server</name>
 	<description>A launcher and stats provider for Toontown Rewritten</description>
 	<properties>
 		<java.version>17</java.version>
