+++ backend/src/main/java/dev/jaczerob/madamchuckle/server/toontown/models/population/District.java
@@ -1,0 +1,12 @@
+package dev.jaczerob.madamchuckle.server.toontown.models.population;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class District {
+    private String name;
+    private int population;
+    private String status;
+}
