+++ backend/src/main/java/dev/jaczerob/madamchuckle/server/toontown/models/releasenotes/ReleaseNotes.java
@@ -1,0 +1,10 @@
+package dev.jaczerob.madamchuckle.server.toontown.models.releasenotes;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+@Data
+@EqualsAndHashCode(callSuper = true)
+public class ReleaseNotes extends ReleaseNotesPartial {
+    private String body;
+}
