+++ backend/src/main/java/dev/jaczerob/madamchuckle/server/toontown/repositories/TotalPopulationRepository.java
@@ -1,0 +1,7 @@
+package dev.jaczerob.madamchuckle.server.toontown.repositories;
+
+import dev.jaczerob.madamchuckle.server.toontown.repositories.entities.TotalPopulationEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TotalPopulationRepository extends JpaRepository<TotalPopulationEntity, Integer> {
+}
