+++ backend/src/main/java/dev/jaczerob/madamchuckle/server/toontown/services/cache/CacheKey.java
@@ -1,0 +1,9 @@
+package dev.jaczerob.madamchuckle.server.toontown.services.cache;
+
+public enum CacheKey {
+    Population,
+    FieldOffices,
+    ReleaseNotes,
+    Status,
+    News
+}
