--- backend/src/main/java/dev/jaczerob/madamchuckle/toontown/models/ToontownObject.java
@@ -1,4 +1,0 @@
-package dev.jaczerob.madamchuckle.toontown.models;
-
-public class ToontownObject {
-}
