--- backend/src/main/java/dev/jaczerob/madamchuckle/toontown/models/releasenotes/ReleaseNotesPartial.java
@@ -1,13 +1,0 @@
-package dev.jaczerob.madamchuckle.toontown.models.releasenotes;
-
-import dev.jaczerob.madamchuckle.toontown.models.ToontownObject;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-@Data
-@EqualsAndHashCode(callSuper = false)
-public class ReleaseNotesPartial extends ToontownObject {
-    private int noteId;
-    private String slug;
-    private String date;
-}
