--- backend/src/main/java/dev/jaczerob/madamchuckle/toontown/services/api/ToontownAPIConfig.java
@@ -1,31 +1,0 @@
-package dev.jaczerob.madamchuckle.toontown.services.api;
-
-import feign.Logger;
-import feign.codec.Decoder;
-import org.springframework.beans.factory.ObjectFactory;
-import org.springframework.boot.autoconfigure.http.HttpMessageConverters;
-import org.springframework.cloud.openfeign.support.SpringDecoder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.MediaType;
-import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
-
-import java.util.Collections;
-
-@Configuration
-public class ToontownAPIConfig {
-    @Bean
-    public Logger.Level feignLoggerLevel() {
-        return Logger.Level.FULL;
-    }
-
-    @Bean
-    @SuppressWarnings("deprecation")
-    public Decoder feignDecoder() {
-        final MappingJackson2HttpMessageConverter converter = new MappingJackson2HttpMessageConverter();
-        converter.setSupportedMediaTypes(Collections.singletonList(MediaType.ALL));
-
-        final ObjectFactory<HttpMessageConverters> objectFactory = () -> new HttpMessageConverters(converter);
-        return new SpringDecoder(objectFactory);
-    }
-}
