--- backend/src/main/java/dev/jaczerob/madamchuckle/toontown/services/cache/CacheKey.java
@@ -1,9 +1,0 @@
-package dev.jaczerob.madamchuckle.toontown.services.cache;
-
-public enum CacheKey {
-    Population,
-    FieldOffices,
-    ReleaseNotes,
-    Status,
-    News
-}
