--- backend/src/test/java/dev/jaczerob/madamchuckle/MadamchuckleApplicationTests.java
@@ -1,11 +1,0 @@
-package dev.jaczerob.madamchuckle;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MadamchuckleApplicationTests {
-	@Test
-	void contextLoads() {
-	}
-}
